Limit POSIX semaphore name to 14 characters
Some *BSD implementations have a limit of 14 characters. Some versions of
FreeBSD and OpenBSD are included. Also from the Oracle man page:

"For maximum portability, name should include no more than 14 characters, but this limit is not enforced."

For now we can limit semaphore name for all implementations for better portability with older
versions as we are using SHA-1 hashing function, so the probability of coincidence is very low.